ai-service/requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
deepface==0.0.79
numpy==1.24.3
Pillow==10.0.1